+++ src/block/blockchain.go
@@ -1,4 +1,4 @@
-package main
+package block
 
 import (
 	"crypto/sha256"
@@ -210,30 +210,4 @@ func (t *Transaction) MarshalJSON() ([]byte, error) {
 		Recipient: t.recipientBlockchainAddress,
 		Value:     t.value,
 	})
-}
-
-/*
-   Initialize
-*/
-
-func init() {
-	log.SetPrefix("Blockchain: ")
-}
-
-func main() {
-	myBlockchainAddress := "my_blockchain_address"
-	blockchain := NewBlockchain(myBlockchainAddress)
-	blockchain.Print()
-
-	blockchain.AddTransaction("A", "B", 1.0)
-	blockchain.Mining()
-	blockchain.Print()
-
-	blockchain.AddTransaction("C", "D", 2.0)
-	blockchain.Mining()
-	blockchain.Print()
-
-	fmt.Printf("my %.1f\n", blockchain.CalculateTotalAmount(myBlockchainAddress))
-	fmt.Printf("C %.1f\n", blockchain.CalculateTotalAmount("A"))
-	fmt.Printf("D %.1f\n", blockchain.CalculateTotalAmount("C"))
 }
