+++ src/main.go
@@ -1,0 +1,12 @@
+package main
+
+import (
+	"log"
+)
+
+func init() {
+	log.SetPrefix("Blockchain: ")
+}
+
+func main() {
+}
